remove lua_lgi and add luajit_lgi
